trying

+++ src/main/java/uk/ac/ed/acp/cw2/DataObjects/DeliveryPath.java
@@ -22,7 +22,7 @@ public class DeliveryPath {
     @Getter
     @Setter
     public static class Delivery {
-        private Long deliveryId;
+        private Integer deliveryId;
         private List<Position> flightPath;
     }
 }
