+++ src/main/java/uk/ac/ed/acp/cw2/DataObjects/ServicePoints.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 @Setter
 public class ServicePoints {
 
-    private Long servicePointId;
+    private Integer servicePointId;
     private List<DroneService> drones;
 
     @Getter
